% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFIMObyGene.R
\name{runFIMObyGene}
\alias{runFIMObyGene}
\title{shiny interface to Kimbie's CISBP FIMO scan TF bed files}
\usage{
runFIMObyGene(ip)
}
\arguments{
\item{ip}{character string with inet ip address}
}
\description{
shiny interface to Kimbie's CISBP FIMO scan TF bed files
}
